requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12

SQLAlchemy==2.0.36
psycopg[binary]==3.2.3

PyYAML==6.0.2
httpx==0.27.2
